setup: move to plugins package
