+++ tests/test_templ.py
@@ -6,7 +6,7 @@
 import inspect
 
 from klpbuild.extractor import Extractor
-from klpbuild.setup import Setup
+from klpbuild.plugins.setup import Setup
 from klpbuild import utils
 from tests.utils import get_file_content
